OA-242 Windows SNMP discovery with incorrect credentials
Major rework of discover_subnet.vbs to nmap all targets and send result as a single POST. Old way was to send each device result separately. This was causing issue's with SNMP discovery when invalid credentials are used. See ticket for details.

Added some extra logging and consistent logging to discovery controller and snmp_helper.

Small syntax error on the html form for result submit (was using translation but it is not loaded for these input forms).

Reworked SNMP helper to use an array of crdentials and loop through them for each attempt to connect. MUCH shorted code and easier to read.